chore: update runtime and dependencies
Upgrade runtime to Node 16 and pg to its latest version to make it
compatible.

Substitute bcrypt with bcryptjs due to installation problems on
Amazon Linux. The library is written in pure Javascript and as such
is less performant than the original, but it's only used in admin
login, so it shouldn't cause any issues.